upd hasValidParameterKDoc()
