+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/ext/provider/KoKDocProviderExt.kt
@@ -7,18 +7,6 @@ import com.lemonappdev.konsist.api.provider.KoPrimaryConstructorProvider
 import com.lemonappdev.konsist.api.provider.KoReturnTypeProvider
 import com.lemonappdev.konsist.api.provider.KoSecondaryConstructorsProvider
 
-/**
- * Whether declaration has a valid KDoc with a PARAM tag.
- *
- * @return `true` if the declaration has a valid KDoc with the PARAM tag, `false` otherwise.
- */
-fun <T> T.hasValidParameterKDoc(): Boolean
-    where T : KoParametersProvider,
-          T : KoKDocProvider = if (parameters.isNotEmpty()) {
-    parameters.count() == kDoc?.paramTags?.count() && parameters.map { it.name } == kDoc?.paramTags?.map { it.value }
-} else {
-    true
-}
 
 /**
  * Whether declaration has a valid KDoc with a PARAM tag.
